Refactor generate_test_report for better library usage.
In the near future we'll be using generate_test_report to process Autotest directories for e-mail digest. The first step here is making sure the code is reusable in library form.

I've factored out the results collection from the report generation and remove dependencies on the options dictionary in favor of individual variables.

I've also made a performance tweak by moving the pruning of invalid directories further up the chain. After the first level, it's safe to assume every directory we're interested in will have a debug directory in it. Moving this up prunes massive trees of possibilities off.

Change-Id: I3a492e83ad6cfa78a033cc2bd2cae22b91cbb6d4

BUG=chromium-os:13496
TEST=Ran a/b comparisons on several results directories.

Review URL: http://codereview.chromium.org/6731024