Added options to run_remote_tests.sh to allow autotests to be profiled.
If profiling is enabled, a header is added to the control file which adds the profiler to the job, with the
specified arguments being passed to the profiler. A footer is added to the control file which deletes the
profiler after the job is complete. The user can specify which autotest profiler they want to use
(defaults to perf) and can specify the arguments to be passed to the initialize function of the profiler.

This is currently the accepted approach for adding profilers to an autotest (see
http://autotest/afe/#tab_id=create_job and the discussion at
https://groups.google.com/a/google.com/group/autotest-users/browse_thread/thread/0e99f3db69e16c7d#).

BUG=chromium-os:15811
TEST=Ran several autotests with and without --profile enabled, with and without arguments passed to the
profiler. Ran a server test with --profile and checked that it fails with an error.

Change-Id: I9be620615c9040d4ea304ed3f5a50aed64d1b4a2
Reviewed-on: http://gerrit.chromium.org/gerrit/1784
Reviewed-by: Ken Mixter <kmixter@chromium.org>
Tested-by: Raymes Khoury <raymes@chromium.org>