Fix case where exception output could be arbitrarily large.
When the multiprocessing module prints exceptions, it uses sys.stderr.write
to write the entire exception in a single operation. If the exception is too
big, this operation can fail with an IOError. So, instead, let's just rely
on the fact that we printed the output already and it doesn't need to be
printed again in the exception.

BUG=chromium-os:14398, chromium-os:16155
TEST=Run sample buildbot run.

Change-Id: I98ecdd30bb48eab2ddced381271b45f7432a4962
Reviewed-on: http://gerrit.chromium.org/gerrit/2155
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: David James <davidjames@chromium.org>