Add parallel test platform_ToolchainOptions
This change modifies the VerifyImage method in AUWorker and children to take
an additional parameter, the string of the test to run. This parameter
defaults to ''; in this case the self.verify_suite string value is used
(which was the behavior previous to this change).

BUG=chromium-os:15521
TEST=Build a test image and run
$bin/cros_au_test_harness --board=x86-mario
--base_image=/home/arkaitzr/chromiumos/src/build/images/x86-mario/latest/chromiumos_test_image.bin
--target_image=/home/arkaitzr/chromiumos/src/build/images/x86-mario/latest/chromiumos_test_image.bin --test_results_root=~/chroot/tmp --no_graphics

and check that it returns

INFO: testPlatformToolchainOptions: Percent passed: 100 vs. Percent required: 100

Change-Id: I8ad6d277f3602a5e360ac8a88a08945ddcfd2442
Reviewed-on: http://gerrit.chromium.org/gerrit/2256
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Arkaitz Ruiz Alvarez <arkaitzr@chromium.org>