Reduce default number of parallel tests to 3.
When I increased the number of tests to run in parallel from 2 to 16, I didn't
see any speed improvement on our bots. (In fact the tests ran a little slower
in my experience, taking about 60 minutes instead of 50.) To reduce memory
pressure I'm going to reduce the number of parallel tests back down to 3 again.

We can wait and see how this affects performance.

BUG=none
TEST=Run test harness.

Change-Id: I0cb6079069bae7e63ce08c7a796806675233c321
Reviewed-on: http://gerrit.chromium.org/gerrit/6039
Reviewed-by: Anush Elangovan <anush@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>