Re-factor and move out generating update payloads to its own program.
The main goal of this is to refactor and redesign out the generating
test payloads out of the au_test_harness into its own module that can
run in parallel without other issues (can also run in parallel without
tripping on the AU timeout bug we've been seeing).  Now that we're moving
to local_archive, I'm removing all the old get update code and using the
local archive.

This is a big change :(

BUG=chromium-os:19479
TEST=Ran ctest a lot.  With /var/www/archive/<build> loaded and without.
With cache set and without (so signed keys).

Change-Id: If2c8970898cb98004bb2e973b466bb4687710080
Reviewed-on: http://gerrit.chromium.org/gerrit/7459
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>