Wait for dev server to start to fix race condition.
When we launch VMs, they may contact the dev server before it has finished
starting and fail as a result. To prevent this, I've added logic to wait
for the dev server to start. This waits up to 3 minutes for the dev server
to print a "Bus STARTED" message to dev_server.log.

BUG=chromium-os:20006, chromium-os:16591
TEST=Simulate slow starting devserver by adding sleep to devserver.
     Trybot run.

Change-Id: I8c770eb085786c3036d7122e346f90883376d3d8
Reviewed-on: http://gerrit.chromium.org/gerrit/7236
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: David James <davidjames@chromium.org>