Fix argument passing issue with embedded spaces in --args flag.
Quoted arguments with embedded spaces provided to the --args flag were
not being properly passed to autoserv. Since quotes are stripped when
the value is passed into run_remote_tests, they need to be added back
before propagating the value to autoserv.

Due to the way autotest splits args, spaces within an argument are still
not allowed; e.g., passing '-a "-i 123"' will result in an args array of
['-i', '123'] instead of ['-i 123']. To avoid this use "-i=123" in such
cases.

BUG=none
TEST=ran run_remote_tests w/ & w/o args.

Change-Id: I3818b5f14802303d21f87427753cd6f9d064281b
Reviewed-on: http://gerrit.chromium.org/gerrit/7755
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Tested-by: Dale Curtis <dalecurtis@chromium.org>