Reserve more RAM for background processes to be more conservative again.
On machines with 12GB RAM or less, we need to limit the number of parallel
VMs to two to prevent the machine from becoming so slow that timeouts occur.

This CL was pushed on Sep 20, and seemed to fix the banner timeout issues.
Then, this CL was accidentally reverted on Oct 3 and the timeouts started
occurring again.

BUG=chromium-os:20305, chromium-os:21351, chromium-os:21389
TEST=Verify this limits jobs to 2 jobs on machines with 12GB RAM. On machines
	 with more RAM we can run more jobs without slowing the tests.

Change-Id: I437ffa97847f6acb693e7b4295f85ef2af1d366a
Reviewed-on: http://gerrit.chromium.org/gerrit/9734
Reviewed-by: Darin Petkov <petkov@chromium.org>
Tested-by: David James <davidjames@chromium.org>