Add ability to ctest to archive nplus1 payloads for future consumption.
This CL adds the ability for ctest to get passed an archive directory
in which we can store nplus1 payloads that are generated.  CBuildbot
or another caller can then be modified to pass in a path to store
these payloads so they can be handed off to other interested parties --
e.g. other test infrastructure.

Also fixed some earlier pylint issues that were introduced since the last
large CL.

BUG=chromium-os:19958, others
TEST=Ran from ctest and cros_generate_update... with / without argument.

Change-Id: Ic0abed6f128c07d3644f61cf924a24793f24fdb5
Reviewed-on: http://gerrit.chromium.org/gerrit/9919
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>