Add unittests to ctest to robustify and fix bug with dirs containing old style archives.
When we turned on the archiving from cbuildbot we found a bug on some of our older builders
where some used to archive and stopped so they have very old builds on them.  This has
led me to the cross point where we really need unittests here so I've added all the cases
I think are important.

I move the logic of max to a reduce function.  Instead of finding the max version in the
local archive folder, we now use a reduce function to reduce the list of files
to the max.  This gives us much more flexiblity to throw out versions that don't match
what we expect (can't do that with max without doing multiple passes over the file list).

BUG=chromium-os:19958
TEST=Ran unittests :)

Change-Id: Ib9428e6ea9a838ba543fc9fc74779cce00b7b9c2
Reviewed-on: http://gerrit.chromium.org/gerrit/10277
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Chris Sosa <sosa@chromium.org>