Fix bug:25622 where we generate non-patched kernel payloads for nplus1 updates.
In this change we move the logic about whether a payload is in a vm into the
individual payload wrapper instead of at a class-level. This way we had it before
we just wrong as we treated all payloads as they were either for a vm or not. This
mean if we wanted to generate both, we never created payloads with the kernel patch.
In order to do this I also had to modify the way we track update payloads between
the test harness and generate_payloads module.

BUG=chromium-os:25622
TEST=Ran with ctest --quick and cros_generate_update_payloads separately
and same args as mario release configuration (generating both nplus1
and vm test payloads).

Change-Id: I8eab3e98f7112a5e8501557257c0993e1b0fbaf6
Reviewed-on: https://gerrit.chromium.org/gerrit/15107
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Chris Sosa <sosa@chromium.org>