Re-land "Fix bug:25622 where we generate non-patched kernel payloads for nplus1 updates.""
In this change we move the logic about whether a payload is in a vm into the
individual payload wrapper instead of at a class-level. This way we had it before
we just wrong as we treated all payloads as they were either for a vm or not. This
mean if we wanted to generate both, we never created payloads with the kernel patch.
In order to do this I also had to modify the way we track update payloads between
the test harness and generate_payloads module.

The original CL had a bug where it assumed VMAUWorker's always had vm_image_path set
which is not true for updates that update from n-1->n->n-1 as the second update
does not start from a vm_image base. I've fixed this by always have for_vm set to
True for any worker performed by a VMAUWorker.

BUG=chromium-os:25622
TEST=Ran with ctest --quick and cros_generate_update_payloads separately
and same args as mario release configuration (generating both nplus1
and vm test payloads).  Validating now without --quick.

Change-Id: Ic8374e3a66a84860937685c9ac3b9b4fa13861e6
Reviewed-on: https://gerrit.chromium.org/gerrit/15246
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Chris Sosa <sosa@chromium.org>