Don't include archive diectories that are missing an image.zip when calculating previous image.
This change allows for a previous archive directory to be devoid of an image.zip
without throwing an exception when generating delta payloads. Sometimes we might
be missing an image.zip directory if the state of the last build was bad i.e. it
was aborted midstream -- archive directory is created but image.zip doesn't finish
getting generated.

I've also updated a lot of the logging/docstring to make clear what's happening
in image_extractor.

BUG=chromium-os:32122
TEST=Unittests + pylint + running with full_suite w/ and w/out image.zip in latest.

Change-Id: I5df2451f24721ed34ff8a0a91c76896083ed9a0c
Reviewed-on: https://gerrit.chromium.org/gerrit/29205
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Chris Sosa <sosa@chromium.org>