Fix ability to specify multiple suite:xxx targets.
This is the second issue we had with associative arrays in bash.
If you do not declare the array with -A, the array is an indexed
array (must be indexed using ints).

This corrects the behavior to use associative arrays and now
multiple suites specified will run more than just running the
last suite control file specified multiple times.

BUG=chromium-os:34359
TEST=run_remote_tests.sh --remote=w.x.y.z --servo --board=daisy
                         firmware_FAFTSetup suite:faft_lv1 suite:faft_lv2

Change-Id: I204c10efb62299f672b03aa1fa9741ec6fcd16b5
Reviewed-on: https://gerrit.chromium.org/gerrit/35283
Reviewed-by: Zdenek Behan <zbehan@chromium.org>
Commit-Ready: Mike Truty <truty@chromium.org>
Tested-by: Mike Truty <truty@chromium.org>