Speed up the check and do not start servod if not necessary
The dut-control command without parameters takes several seconds to
complete. But to determine if servod is running it is enough to
retrieve a single value, and retrieving most of the values is very fast.

Also, if servod is already running on the system, we should not be
just ignoring it and proceeding with starting our own instance. The
suggested way of handling it is not to start servod if it is already
running (as opposed to interrupting the test), but warn the user about
the situation.

BUG=chrome-os-partner:15610
TEST=manual

  . have a servod running on a system, start run_remote_tests.sh on
    the same system, observe

    WARNING : --servo option ignored (servod already running!)

   showing up in the log

Change-Id: I55fedee8a99958a748362cf1250eafcc78afc9a3
Reviewed-on: https://gerrit.chromium.org/gerrit/37051
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Ready: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>