Fixed bottle views folder path and mimetype.
1. Adding the views folder to bottle.TEMPLATE_PATH
which allows us to remove the need for changing
into the same directory as the script. The dir
change was confusing python when the script was
called in certain scenarios, like from relative
paths.

Also, removed some functions that are no longer
needed.

2. Setting correct mimetype for html files. We
need to have the default be text/plain because
most log files are not detected correctly and they
are rendered ugly in the browser.

BUG=None
TEST=manual run

Change-Id: Id83a2018062566b6609bd2bdb8f7db90a2afdf6f
Reviewed-on: https://gerrit.chromium.org/gerrit/37471
Tested-by: Yusuf Mohsinally <mohsinally@google.com>
Reviewed-by: Mike Truty <truty@chromium.org>
Commit-Ready: Yusuf Mohsinally <mohsinally@google.com>