crostestutils: use "--copy" option when starting VM
We sometimes see failures where we are unable to log in to the VM
over ssh. In these cases, we can't fetch log files for the VM,
because we normally do that via ssh.

We could fall back to mounting the VM's filesystem image, if not
for the fact that we use "--snapshot", which saves deltas to
temporary files.

Switch from "--snapshot" to "--copy", so that we have a VM filesystem
image that includes log files from whatever happened during the VM's
lifetime.

CQ-DEPEND=CL:63111
BUG=chromium:212084
TEST=trybot on x86-generic-paladin, amd64-generic-paladin, daisy-paladin

Change-Id: If4f8da2c548d75aae56dd204d14d7f99af096963
Reviewed-on: https://gerrit.chromium.org/gerrit/63114
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>