cros_run_vm_test: turn down ssh verbosity in VMTest
The ssh verbosity was turned up to 3 in VMTest to help trace the
referenced bug ("banner exchange" flake). The issue seems to be mostly
mitigated now, and from an ssh log standpoint we already know all we
need to know. This CL turns the verbosity back down, to reduce the
amount of noise in the logs.

BUG=chromium:254166
TEST=None

Change-Id: I88fb93acbc4e80ce9bb6799fc23687a9d2e8f5c9
Reviewed-on: https://chromium-review.googlesource.com/172945
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>