Allow passing target image name to ImageExtractor
When generating the delta payload, we use ImageExtractor to extract
the latest image and use the image as a base to generate the delta
payload. Because ImageExtractor always assumes that the image to
extract is a test image ('chromiumos_test_image.bin'), we are unable
to generate the delta payload for non-test images. This needs to be
changed because there are builders (e.g. beaglebone) which do not
produces test images. This CL passes the target image name to
ImageExtractor so that a correct type of image can be used.

BUG=chromium:326632
TEST=unittest

Change-Id: Ibb347b4cccbf55079bc719200278dac74e0ab87f
Reviewed-on: https://chromium-review.googlesource.com/179085
Reviewed-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Yu-Ju Hong <yjhong@chromium.org>
Commit-Queue: Yu-Ju Hong <yjhong@chromium.org>
Tested-by: Yu-Ju Hong <yjhong@chromium.org>