Create a script to run the wifi_release suite.
Due to http://crbug.com/241416 suites cannot be run from the AFE.
This script will detect which devices are ready and queue the suite
to run.  Since there are a limited amount of machines, boards and
put into groups and only one in any group is required.  The script
will notify the user if no boards in a given group are available.

TEST=Manual
BUG=None

Change-Id: I71413f65264af1326a2f998a28258ac050e97b49
Reviewed-on: https://chromium-review.googlesource.com/183573
Reviewed-by: Kris Rambish <krisr@chromium.org>
Tested-by: Kris Rambish <krisr@chromium.org>