If an Zip of a previous image is corrupt, don't 'find' it.
We were failing our updates tests against a VM image from the previous build
if that image was corrupt. This assumes that the image would not have been
zipped if it was corrupt, but that the zipfile itself is corrupt.

So... ignore any zipped images in which the zip itself is corrupt.

BUG=chromium:293515
TEST=lint + unittests. Unable to replicate original failure (crbug.com/347259)

Change-Id: I0db7eca9f004c8d68fd12b8ca0456c478a36e605
Reviewed-on: https://chromium-review.googlesource.com/188084
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>