Increase logging of check_ethernet.hook.
Prior to this change, the only message check_ethernet.hook wrote
to /var/log/messages was to report reboot.  Other interesting events
(such as attempting to restart the network) went unreported.  This
adds logging messages to check_ethernet.hook to report the first
time ping fails to reach the gateway, and again before attempting to
restart ethernet interfaces.

The motivation for this change is to increase visibility during
unexplained network outages on the DUT.

BUG=chromium:354944
TEST=run check_ethernet.hook on a DUT with and without a network

Change-Id: I77dfbfe51aee118c7802726379216948be45620f
Reviewed-on: https://chromium-review.googlesource.com/197529
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Prashanth B <beeps@chromium.org>