ctest: Change 'quick' flag to 'quick_update'.
The --quick flag of ctest would run SimplyVerifyAndUpdateTest as well as
SimplyVerifyTest tests, in addition to generating payloads only for N->N update.

We would like to keep these two options separate: The case where the image is
tested directly (without running auto-update) is already covered by the
--verify_only flag.

This CL renames --quick to --quick_update, and restricts the tests ran to just
the SimpleVerifyAndUpdateTest. It is still possible to get the effect of the old
--quick flag by running ctest twice, once with --verify_only and --quick_update
each.

BUG=chromium:430529
TEST=`ctest --quick_update --board ${BOARD} --type=vm` only runs
     SimplyUpdateAndVerifyTest.
CQ-DEPEND=CL:227739

Change-Id: I0d9fc4bb65cff9a5aa4025213a9893966da770ba
Reviewed-on: https://chromium-review.googlesource.com/227697
Reviewed-by: David James <davidjames@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>