Update browsertest_status.py to work with ToT and branch builders.
This update allows the browsertest_status.py to work with both ToT and
Branch builders. To accommodate this, it adds two command line flags
to specify:
- Build Host
- Build Project
- Chromium Version

In an unrelated change, it adds a flag to enable creating a report
file, which is now disabled by default.

BUG=chromium:375395
TEST=Put a tests file in the same directory as the script, that lists
the browsertests you want to have a status/results report. Then run:
$ ./browsertest_status.py --tests_file=tests

Change-Id: I7fd5a1516aaacc2422ea6ff784400843ea6bbe23
Reviewed-on: https://chromium-review.googlesource.com/262432
Reviewed-by: Scott Cunningham <scunningham@chromium.org>
Commit-Queue: Scott Cunningham <scunningham@chromium.org>
Tested-by: Scott Cunningham <scunningham@chromium.org>