crostestutils: Add a GCE Python API based utility
GCE Python API is now preferred to wrapping the gcloud command line tool. It is
more programmatically robust and is already available on most build slaves. This
utility is added in particular for use in gce_au_worker, but may be extended to
replace the current gcloud.py in chromite/compute, that wrapps the gcloud
command line.

BUG=brillo:1196
TEST=Unit test implemented in CL:299694. It contains project credentials so is
only included for demonstration. Will not be submitted.

Change-Id: Iab932a68bd247fa848c520323ca9c8286edbf99d
Reviewed-on: https://chromium-review.googlesource.com/300202
Commit-Ready: Daniel Wang <wonderfly@google.com>
Tested-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Fang Deng <fdeng@chromium.org>