crostestutils: Move lib/gce.py to chromite
lib/gce.py is getting more attraction, so better move it to chromite.lib, which
is the best place for common libraries for cross-repo usage. One immediate use
of it would be the server host for GCE that we are adding to Autotest.

BUG=b:25291982
TEST=au_test_harness/gce_au_worker_unittest.py and trybot against lakitu-pre-cq
CQ-DEPEND=CL:311833

Change-Id: Ic5b773ebb4bbbcc425a941e301d3699e48a70bf4
Reviewed-on: https://chromium-review.googlesource.com/311844
Commit-Ready: Daniel Wang <wonderfly@google.com>
Tested-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Fang Deng <fdeng@chromium.org>