gce_au_worker: Avoid using colon in file names
The `test_that` script uses colon as a seperator in autotest suite names, e.g.,
`test_that -b lakitu 1.1.1.1 suite:gce-smoke`. We used to use the full suite
name as file name when saving test results, and it confused a piece of chromite
code (chromite.commands.ListFailedTests, which uses a regular expression to
match file names and colon is not deemed as a valid character) that parses test
logs for failed tests.

A sample failure can be found here (notice the 'IndexError' at the bottom of
the log):
https://uberchromegw.corp.google.com/i/chromeos/builders/lakitu%20canary/builds/712/steps/GCETest%20%28attempt%202%29/logs/stdio

While we can change the chromite code to include colon, I think it's better to
not use colon in file names at all, as it is commonly used as seperator in
things like $PATH and URL.

This CL also tweaks VerifyImage to print out the summarized test report in case
of a test failure, and includes some style smashes to better conform with the
style guide.

BUG=b:25849966
TEST=Unit test. Also ran a trybot against lakitu-pre-cq, with an intentional
    test failure and verified that failed test was correctly linked in the
    buildbot log, and the 'IndexError' was gone.

Change-Id: Ie4b3b002c95b81f30c4c63799b77819651050757
Reviewed-on: https://chromium-review.googlesource.com/317781
Commit-Ready: Daniel Wang <wonderfly@google.com>
Tested-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Amey Deshpande <ameyd@google.com>